Move out the layout
