+++ js/src/components/lineage/LineageView.tsx
@@ -4,6 +4,7 @@ import {
   LineageGraphNode,
   cleanUpNodes,
   highlightNodes,
+  layout,
   selectDownstream,
   selectNode,
   selectNodes,
@@ -41,7 +42,6 @@ import ReactFlow, {
   ControlButton,
   useReactFlow,
 } from "reactflow";
-import dagre from "dagre";
 import "reactflow/dist/style.css";
 import { GraphNode } from "./GraphNode";
 import GraphEdge from "./GraphEdge";
@@ -77,40 +77,6 @@ export interface LineageViewProps {
   viewMode?: "changed_models" | "all";
   filterNodes?: (key: string, node: LineageGraphNode) => boolean;
 }
-
-const layout = (nodes: Node[], edges: Edge[], direction = "LR") => {
-  const dagreGraph = new dagre.graphlib.Graph();
-  dagreGraph.setDefaultEdgeLabel(() => ({}));
-
-  const nodeWidth = 300;
-  const nodeHeight = 36;
-
-  const isHorizontal = direction === "LR";
-  dagreGraph.setGraph({ rankdir: direction });
-
-  nodes.forEach((node) => {
-    dagreGraph.setNode(node.id, { width: nodeWidth, height: nodeHeight });
-  });
-
-  edges.forEach((edge) => {
-    dagreGraph.setEdge(edge.source, edge.target);
-  });
-
-  dagre.layout(dagreGraph);
-
-  nodes.forEach((node) => {
-    const nodeWithPosition = dagreGraph.node(node.id);
-
-    // We are shifting the dagre node position (anchor=center center) to the top left
-    // so it matches the React Flow node anchor point (top left).
-    node.position = {
-      x: nodeWithPosition.x - nodeWidth / 2,
-      y: nodeWithPosition.y - nodeHeight / 2,
-    };
-
-    return node;
-  });
-};
 
 const nodeTypes = {
   customNode: GraphNode,
