+++ js/src/components/lineage/lineage.ts
@@ -1,7 +1,7 @@
 import { Node, Edge, Position } from "reactflow";
 import { getNeighborSet, union } from "./graph";
 import { Run } from "@/lib/api/types";
-
+import dagre from "dagre";
 /**
  * The data from the API
  */
@@ -300,6 +300,40 @@ export function toReactflow(
 
   return [nodes, edges];
 }
+
+export const layout = (nodes: Node[], edges: Edge[], direction = "LR") => {
+  const dagreGraph = new dagre.graphlib.Graph();
+  dagreGraph.setDefaultEdgeLabel(() => ({}));
+
+  const nodeWidth = 300;
+  const nodeHeight = 36;
+
+  const isHorizontal = direction === "LR";
+  dagreGraph.setGraph({ rankdir: direction });
+
+  nodes.forEach((node) => {
+    dagreGraph.setNode(node.id, { width: nodeWidth, height: nodeHeight });
+  });
+
+  edges.forEach((edge) => {
+    dagreGraph.setEdge(edge.source, edge.target);
+  });
+
+  dagre.layout(dagreGraph);
+
+  nodes.forEach((node) => {
+    const nodeWithPosition = dagreGraph.node(node.id);
+
+    // We are shifting the dagre node position (anchor=center center) to the top left
+    // so it matches the React Flow node anchor point (top left).
+    node.position = {
+      x: nodeWithPosition.x - nodeWidth / 2,
+      y: nodeWithPosition.y - nodeHeight / 2,
+    };
+
+    return node;
+  });
+};
 
 export function highlightNodes(
   nodesIds: string[],
